feat(card&deck) : DIVIDE - Card, Deck, Denomination, Shape class
