+++ src/main/kotlin/blackjack/Card.kt
@@ -1,0 +1,6 @@
+package blackjack
+
+class Card(private val denominationToShape: Pair<Denomination, Shape>) {
+
+    override fun toString(): String = "${denominationToShape.first.symbol}${denominationToShape.second.symbol}"
+}
