+++ src/main/kotlin/blackjack/Shape.kt
@@ -1,0 +1,8 @@
+package blackjack
+
+enum class Shape(val symbol: String) {
+    SPADE("스페이드"),
+    HEART("하트"),
+    DIAMOND("다이아"),
+    CLUB("클로버")
+}
